requirements.txt
opencv-python
pyyaml
psycopg2-binary
flask
fastapi
uvicorn
pydantic
numpy
onnxruntime
supervision
insightface
mediapipe
requests
mcp
